+++ src/lib/Driver.Redis.ts
@@ -1,26 +1,26 @@
 /**
- *  Copyright 2018 Angus.Fenying <fenying@litert.org>
+ * Copyright 2020 Angus.Fenying <fenying@litert.org>
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *   https://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
-import * as Mutex from "@litert/mutex";
-import * as Redis from "@litert/redis";
+import * as Mutex from '@litert/mutex';
+import * as Redis from '@litert/redis';
 import {
     IKeyBuilder,
     DEFAULT_NAMESPACE,
     DEFAULT_KEY_BUILDER
-} from "./Common";
+} from './Common';
 
 const LUA_LOCK: string = `if 1 == redis.call('setnx', KEYS[1], ARGV[1]) then
 
@@ -55,7 +55,7 @@ return 0`;
 class RedisDriver
 implements Mutex.IDriver {
 
-    private _redis: Redis.RedisClient;
+    private _redis: Redis.ICommandClient;
 
     private _ns: string;
 
@@ -68,7 +68,7 @@ implements Mutex.IDriver {
     private _luaCheckId!: string;
 
     public constructor(
-        redis: Redis.RedisClient,
+        redis: Redis.ICommandClient,
         namespace: string = DEFAULT_NAMESPACE,
         keyBuilder: IKeyBuilder = DEFAULT_KEY_BUILDER
     ) {
@@ -77,9 +77,9 @@ implements Mutex.IDriver {
         this._ns = namespace;
         this._buildKey = keyBuilder;
 
-        this._luaCheckId = "cc5e3d7f05b5e6c28c556242d5a24eece2a1ae2a";
-        this._luaLockId = "3a971c811a10347bbdee17970f8b6ed3cb6fd7f7";
-        this._luaUnlockId = "d03dcdcd1e51e59673c900e403199e9544f77b26";
+        this._luaCheckId = 'cc5e3d7f05b5e6c28c556242d5a24eece2a1ae2a';
+        this._luaLockId = '3a971c811a10347bbdee17970f8b6ed3cb6fd7f7';
+        this._luaUnlockId = 'd03dcdcd1e51e59673c900e403199e9544f77b26';
     }
 
     private async _registerLuaScripts(): Promise<void> {
@@ -124,7 +124,7 @@ implements Mutex.IDriver {
                     return false;
                 }
 
-                if (e.message && e.message.startsWith("NOSCRIPT")) {
+                if (e.metadata?.message?.startsWith('NOSCRIPT')) {
 
                     await this._registerLuaScripts();
                 }
@@ -161,7 +161,7 @@ implements Mutex.IDriver {
                     return false;
                 }
 
-                if (e.message && e.message.startsWith("NOSCRIPT")) {
+                if (e.metadata?.message?.startsWith('NOSCRIPT')) {
 
                     await this._registerLuaScripts();
                 }
@@ -198,7 +198,7 @@ implements Mutex.IDriver {
                     return false;
                 }
 
-                if (e.message && e.message.startsWith("NOSCRIPT")) {
+                if (e.metadata?.message?.startsWith('NOSCRIPT')) {
 
                     await this._registerLuaScripts();
                 }
@@ -216,7 +216,7 @@ implements Mutex.IDriver {
  * Create a driver for Redis.
  */
 export function createRedisDriver(
-    redis: Redis.RedisClient,
+    redis: Redis.ICommandClient,
     namespace: string = DEFAULT_NAMESPACE,
     keyBuilder: IKeyBuilder = DEFAULT_KEY_BUILDER
 ): Mutex.IDriver {