+++ CHANGES.md
@@ -1,3 +1,7 @@
 # Changes Logs
 
-## v0.1.0
+## v0.2.0
+
+- fix(redis): adapt to the new version of @litert/redis.
+- config(deps): updated the deps.
+- config(lint): replaced tslint with eslint.